Fixed tokens. Improved profile.
Tokens were not being loaded propery into axios on refresh of the 'Projects' set of pages and as such no projects were visible in the sidebar. Added some basic user data (because I'm not certain that having all data be visible and stored in localStorage would be safe, but in hindsight it's probably ok to make it available to the user considering they logged in - either way, this is enough data for the functionality). The profiles tab was simplified by removing unused buttons/links and now greets the user with their name and shows their role.