Collective statistics
Created a new page for the collective statistics. It shows a chart with a line and an area around it. The line represents average data accross all recorded reactions, and the area represents one standard deviation from that average. Still need to add the appropriate headers and similar housekeeping stuff, but the meaty part is considreded finished.